backend/pyproject.toml
[project]
name = "ultr-erp-backend"
version = "0.1.0"
description = "UltrERP FastAPI backend"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
	"alembic>=1.14.0",
	"asyncpg>=0.30.0",
	"fastapi>=0.115.0",
	"greenlet>=3.1.1",
	"pydantic-settings>=2.6.1",
	"python-multipart>=0.0.17",
	"redis>=5.2.1",
	"sqlalchemy>=2.0.39",
	"uvicorn[standard]>=0.34.0",
	"weasyprint>=63.0"
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.uv]
package = false

[dependency-groups]
dev = [
	"httpx>=0.28.1",
	"pytest>=8.3.5",
	"pytest-asyncio>=0.26.0",
	"ruff>=0.11.2"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
